updated UI for Event Views
