+++ Olympsis/Components/Post/GroupFeed.swift
@@ -167,7 +167,7 @@ struct GroupFeed: View {
                             })
                             
                             if let event = groupEvents.first {
-                                EventView(event: event)
+                                EventListItemView(event: event)
                             }
                         }.padding(.horizontal)
                             .padding(.vertical)
